large batch of Dehydra static code analysis fixes (mostly passing arguments by const reference and checking LibError return codes)
git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@6214 3db68df2-c116-0410-a063-a993310a9797

+++ source/graphics/MapReader.cpp
@@ -170,8 +170,8 @@ int CMapReader::UnpackTerrain()
 	}
 
 	// unpack tile data [3ms]
-	size_t tilesPerSide = m_MapSize*PATCH_SIZE;
-	m_Tiles.resize(SQR(tilesPerSide));
+	ssize_t tilesPerSide = m_MapSize*PATCH_SIZE;
+	m_Tiles.resize(size_t(SQR(tilesPerSide)));
 	unpacker.UnpackRaw(&m_Tiles[0], sizeof(STileDesc)*m_Tiles.size());
 
 	// reset generator state.
